\name{demovalgrindnonfail-package}
\alias{demovalgrindnonfail-package}
\alias{demovalgrindnonfail}
\docType{package}
\title{\packageTitle{demovalgrindnonfail}}
\description{\packageDescription{demovalgrindnonfail}}
\section{Package Content}{\packageIndices{demovalgrindnonfail}}
\author{\packageAuthor{demovalgrindnonfail}}
\section{Maintainer}{\packageMaintainer{demovalgrindnonfail}}
\keyword{package}
